+++ NEWS.md
@@ -36,7 +36,7 @@ sparse_model_matrix(mpg ~ i(vs) | gear^cyl, data = mtcars, type = c("rhs", "fixe
 
 - the functions `coefplot` and `iplot` now accept models passsed via `...`, aligning their design to the one of `etable`
 
-- the functions `coefplot` and `iplot` now accept the argument `vcov`
+- the functions `coefplot` and `iplot` now accept the argument `vcov`. The argument `vcov` can be a list of valid `vcov` values, in which case it will be recycled across the models.
 
 ## New vignette and data set